+++ addon/components/generic-input-field.js
@@ -1,6 +1,6 @@
 import Ember from 'ember';
 import assign from '../utils/deep-assign';
-import layout from '../templates/components/root-component';
+import layout from '../templates/components/generic-input-field';
 const { A, Component, computed, get } = Ember;
 
 export default Component.extend({
