+++ addon/templates/components/generic-input-field.hbs
@@ -2,7 +2,7 @@
 
   {{#if resolvedAll.length}}
     <div class='generic-field'>
-      {{generic-field
+      {{recursive-input
         content=(get-parents resolvedAll)
         tree=myTree
