--- app/components/generic-field.js
@@ -1,1 +1,0 @@
-export { default } from 'generic-input-field/components/generic-field';
