+++ app/components/generic-input-field.js
@@ -1,0 +1,1 @@
+export { default } from 'generic-input-field/components/generic-input-field';
