+++ app/components/recursive-input.js
@@ -1,0 +1,1 @@
+export { default } from 'generic-input-field/components/recursive-input';
