--- app/components/root-component.js
@@ -1,1 +1,0 @@
-export { default } from 'generic-input-field/components/root-component';
