+++ tests/dummy/app/templates/application.hbs
@@ -1,4 +1,4 @@
-{{root-component
+{{generic-input-field
   all=myAll
   input=myInput
